added separate module for stocks
